+++ src/geometry/geometry_operations.rs
@@ -1,7 +1,7 @@
 use nalgebra::base::*;
 
 /// Checks if the 4 given points are planar
-pub fn is_flat(pts: [Vector3<f32>; 4], eps: Option<f32>) -> bool {
+pub fn is_flat(pts: [Vector3<f64>; 4], eps: Option<f64>) -> bool {
     let eps_val = eps.unwrap_or(0.00001);
     let vec_3_0n = (pts[0] - pts[3]).normalize();
     let vec_3_1n = (pts[1] - pts[3]).normalize();
@@ -20,7 +20,7 @@ pub fn is_flat(pts: [Vector3<f32>; 4], eps: Option<f32>) -> bool {
 }
 
 /// Computes sphere center associated to the 4 given points
-pub fn sphere_center(pts: [Vector3<f32>; 4]) -> Option<Vector3<f32>> {
+pub fn sphere_center(pts: [Vector3<f64>; 4]) -> Option<Vector3<f64>> {
     let vec_0_1 = pts[1] - pts[0];
     let vec_1_2 = pts[2] - pts[1];
     let vec_2_0 = pts[0] - pts[2];
@@ -58,7 +58,7 @@ pub fn sphere_center(pts: [Vector3<f32>; 4]) -> Option<Vector3<f32>> {
 }
 
 /// Computes circle center associated to the 3 given points
-pub fn circle_center(pts: [Vector3<f32>; 3]) -> Option<Vector3<f32>> {
+pub fn circle_center(pts: [Vector3<f64>; 3]) -> Option<Vector3<f64>> {
     let vec_0_1 = pts[1] - pts[0];
     let vec_1_2 = pts[2] - pts[1];
     let vec_2_0 = pts[0] - pts[2];
@@ -89,7 +89,7 @@ pub fn circle_center(pts: [Vector3<f32>; 3]) -> Option<Vector3<f32>> {
 /// Computes sphere center and radius associated to the 4 given points
 ///
 /// If the 4 points are planar, a circle center is computed instead
-pub fn center_and_radius(pts: [Vector3<f32>; 4], eps: Option<f32>) -> Option<(Vector3<f32>, f32)> {
+pub fn center_and_radius(pts: [Vector3<f64>; 4], eps: Option<f64>) -> Option<(Vector3<f64>, f64)> {
     let center = if is_flat(pts, eps) {
         circle_center([pts[0], pts[1], pts[2]])
     } else {
